Use mocha directly for testing
