+++ test/acceptance/blank_description_test.js
@@ -1,4 +1,3 @@
-testSupport.requireICAL();
 
 suite('ics - blank description', function() {
   var icsData;
