+++ test/acceptance/daily_recuring_test.js
@@ -1,5 +1,3 @@
-testSupport.requireICAL();
-
 suite('ics - blank description', function() {
   var icsData;
