+++ test/acceptance/forced_types_test.js
@@ -1,5 +1,3 @@
-testSupport.requireICAL();
-
 suite('ics test', function() {
   var icsData;
