+++ test/acceptance/utc_negative_zero_test.js
@@ -1,5 +1,3 @@
-testSupport.requireICAL();
-
 suite('ics - negative zero', function() {
   var icsData;
